xfail test_rewrite_assert tests on Python 3.9 and above
See also https://github.com/pytorch/torchdynamo/issues/1922


[ghstack-poisoned]
